V.2

+++ app/Http/Controllers/my_api.php
@@ -16,7 +16,7 @@ class my_api extends Controller
     public function index()
     {
 
-         return articleresource::collection(Article::with('user')->paginate(25));
+         return articleresource::collection(Article::with('user')->latest()->paginate(25));
         //$articles=Article::paginate(15);
         //print_r($articles);
 
@@ -128,9 +128,16 @@ Article::where('id',$id)->update($data);
         //dleting from article table
 
         $article=Article::FindOrFail($id);
-        if($article->delete())
-        {
-        return new articleresource($article);
+        $logged_user=auth('api')->user()->id;
+        $del_article=$article->user_id;
+        if($logged_user==$del_article){
+          if($article->delete())
+          {
+          return 'deletion successfully';
         }
+        }else{
+          return "you dont have enough permission contact your system administrator";
+        }
+
     }
 }
